requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
urllib3>=2.0.7

# MCP framework
mcp>=0.5.0

# For browser automation (optional)
selenium>=4.15.0
webdriver-manager>=4.0.1

# Utility libraries
pydantic>=2.5.0
